Use BatchPoints for writing from client library
