+++ client/influxdb_test.go
@@ -111,8 +111,8 @@ func TestClient_Write(t *testing.T) {
 		t.Fatalf("unexpected error.  expected %v, actual %v", nil, err)
 	}
 
-	write := client.Write{}
-	_, err = c.Write(write)
+	bp := client.BatchPoints{}
+	_, err = c.Write(bp)
 	if err != nil {
 		t.Fatalf("unexpected error.  expected %v, actual %v", nil, err)
 	}
@@ -172,8 +172,8 @@ func TestClient_UserAgent(t *testing.T) {
 		}
 
 		receivedUserAgent = ""
-		write := client.Write{}
-		_, err = c.Write(write)
+		bp := client.BatchPoints{}
+		_, err = c.Write(bp)
 		if err != nil {
 			t.Fatalf("unexpected error.  expected %v, actual %v", nil, err)
 		}
@@ -339,4 +339,43 @@ func emptyTestServer() *httptest.Server {
 		w.Header().Set("X-Influxdb-Version", "x.x")
 		return
 	}))
+}
+
+// Ensure that data with epoch timestamps can be decoded.
+func TestBatchPoints_Normal(t *testing.T) {
+	var bp client.BatchPoints
+	data := []byte(`
+{                                                                                                                                                       
+    "database": "foo",                                                                                                                                    
+    "retentionPolicy": "bar",                                                                                                                              
+    "points": [                                                                                                                                            
+        {                                                                                                                                                   
+            "name": "cpu",                                                                                                                                   
+            "tags": {                                                                                                                                         
+                "host": "server01"                                                                                                                            
+            },
+            "timestamp": 14244733039069373,
+            "precision": "n",
+            "values": {
+                    "value": 4541770385657154000
+            }
+        },
+        {
+            "name": "cpu",
+             "tags": {
+                "host": "server01"
+            },
+            "timestamp": 14244733039069380,
+            "precision": "n",
+            "values": {
+                    "value": 7199311900554737000
+            }
+        }
+    ]
+}
+`)
+
+	if err := json.Unmarshal(data, &bp); err != nil {
+		t.Errorf("failed to unmarshal nanosecond data: %s", err.Error())
+	}
 }
