+++ server.go
@@ -19,6 +19,7 @@ import (
 	"sync"
 	"time"
 
+	"github.com/influxdb/influxdb/client"
 	"github.com/influxdb/influxdb/influxql"
 	"github.com/influxdb/influxdb/messaging"
 	"golang.org/x/crypto/bcrypt"
@@ -1365,6 +1366,45 @@ type Point struct {
 	Tags      map[string]string
 	Timestamp time.Time
 	Fields    map[string]interface{}
+}
+
+// NormalizeBatchPoints returns a slice of Points, created by populating individual
+// points within the batch, which do not have timestamps or tags, with the top-level
+// values.
+func NormalizeBatchPoints(bp client.BatchPoints) ([]Point, error) {
+	points := []Point{}
+	for _, p := range bp.Points {
+		if p.Timestamp.Time().IsZero() {
+			if bp.Timestamp.IsZero() {
+				p.Timestamp = client.Timestamp(time.Now())
+			} else {
+				p.Timestamp = client.Timestamp(bp.Timestamp)
+			}
+		}
+		if p.Precision == "" && bp.Precision != "" {
+			p.Precision = bp.Precision
+		}
+		p.Timestamp = client.Timestamp(client.SetPrecision(p.Timestamp.Time(), p.Precision))
+		if len(bp.Tags) > 0 {
+			if p.Tags == nil {
+				p.Tags = make(map[string]string)
+			}
+			for k := range bp.Tags {
+				if p.Tags[k] == "" {
+					p.Tags[k] = bp.Tags[k]
+				}
+			}
+		}
+		// Need to convert from a client.Point to a influxdb.Point
+		points = append(points, Point{
+			Name:      p.Name,
+			Tags:      p.Tags,
+			Timestamp: p.Timestamp.Time(),
+			Fields:    p.Fields,
+		})
+	}
+
+	return points, nil
 }
 
 // WriteSeries writes series data to the database.
